just getting it closer to working, lots of changes to make sure it compiles as well

+++ data/bout.hpp
@@ -1,9 +1,10 @@
 #pragma once
 
 #include <constants.hpp>
+#include <stdint.h>
 #include <data/lick.hpp>
 
 struct bout {
     lick licks[constants::max_bout_size];
-    size_t count = 0; // Tracks how many licks are actually valid in this array
+    uint32_t count = 0; // Tracks how many licks are actually valid in this array
 };
